refactor: move fetch games function to separate directory
